+++ changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 
 All notable changes to `SiBex` will be documented in this file.
+## Version 0.0.3
+
+### Added
+- Added new function to get all contacts from a list.
+- Added translator of SIB model to object or collection for better processing.
 
 ## Version 0.0.2